Move code outside main test block
